requirements_dev.txt
nonebot2[fastapi]>=2.0.0
nonebot-adapter-onebot>=2.0.0
aiohttp>=3.8.0
jinja2>=3.0.0
uvicorn>=0.20.0
